requirements.txt
tensorflow>=2.12
streamlit
numpy
Pillow